Allow disutils to detect implementation of linear algebra
Currently, the libraries for linear algebra are hard coded and
constrained. This makes the installation to fail in several
environments, even when available BLAS/LAPACK libraries
are present. This happens frequently with MKL. Since the
interface of these routines is always the same, it is worth it
to have a dynamical detection of the library, which is done
through numpy directly.